Added Functionality -> Verify Admin. Now only admin can make changes in inventory. If admin not verfied so user can only visit inventory  user cannot be made changes in inventory.
Used a global varible 'check' in index.js
initialised check = 0 which means admin not verified.
After verification check will be 1.
so if check = 0 it means admin not verified.
if check = 1 it means admin verified.

when user click on any button of side-bar except
'Visit Inventory' so if user is not verifed means if 'check' is 0 so response will sent to template verify_admin.ejs. In this template  user will fill username/email and password.

From verify_admin.ejs template post request will sent to '/home/verify'. Inside this request we will extract original username/email and password of admin from database and we will match the information filled by user with the admin data.

if username/ email and password matches it means  admin verified so we will make 'check' = 1 and redirect to home page.

Otherwise[ if admin not verified]  render to not_verifed.ejs template. In this template there will be message of verification unsucessful. From this page you can Retry the verification as well you can go to home page.