Added 'Admin Profile page' also  'Sign Out' functionality. Now you can visit to Admin profile also admin can signout itself.
Admin Profile Page -: On clicking 'Admin' button of navbar, user will redirect either to the admin profile page or verification page.

Case1: if admin is already verifed so request will redirect to admin profile means admin_page.ejs template.

Case2: if admin is not verified so request will render to verify_admin.ejs template to verify admin.

Signout Button -: Inside the admin profile page there is a 'Signout' button on clicking it admin can signout itself.
On clicking 'Signout' button 'get request' will sent to 'home/signout'. In this request varible 'check' will change from 1 to 0.
means make check = 0.

'check' is a global varible. which indicates verification of admin means it checks is admin verified or not.

if check = 0 it means admin is not verified.
if check = 1 it means admin is verified.